% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_legacy.R
\name{equals_legacy}
\alias{equals_legacy}
\title{equals_legacy}
\usage{
equals_legacy(expected, label = NULL, ...)
}
\arguments{
\item{expected}{expected}

\item{label}{label}

\item{...}{rest}
}
\description{
equals_legacy
}

