% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{exercise}
\alias{exercise}
\title{Take an instructor-provided exercise with SocraticSwirl}
\usage{
exercise(question)
}
\arguments{
\item{question}{Which quiz question to take; provided by instructor}
}
\description{
This is to be called after \link{\code{socratic_swirl}} is used to set up
a SocraticSwirl session.
}

