% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{install_course_socratic_swirl}
\alias{install_course_socratic_swirl}
\title{install a course from the Socratic Swirl server}
\usage{
install_course_socratic_swirl(course)
}
\arguments{
\item{course}{Course title}
}
\description{
Given the title of a course, install it from the server
}

