% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{notify_socratic_swirl}
\alias{notify_socratic_swirl}
\title{Given a Swirl environment, update SocraticSwirl server}
\usage{
notify_socratic_swirl(e, correct = TRUE)
}
\arguments{
\item{e}{Swirl environment, containing info on the current Swirl session}

\item{correct}{whether the answer was correct}
}
\value{
boolean describing whether it uploaded the Socratic Swirl results
}
\description{
Given a Swirl environment, update SocraticSwirl server
}

