% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{Parse_retrieve}
\alias{Parse_retrieve}
\title{retrieve one or more objects from Parse}
\usage{
Parse_retrieve(class_name, object_id, ...)
}
\arguments{
\item{class_name}{}

\item{object_id}{if provided, a specific object ID to retrieve}

\item{...}{fields to query by (not yet implemented)}
}
\description{
retrieve one or more objects from Parse
}

