% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{socratic_swirl}
\alias{socratic_swirl}
\title{set up SocraticSwirl in this session}
\usage{
socratic_swirl(course, lesson, student, instructor = "mcahn",
  instance = "prod")
}
\arguments{
\item{course}{course name}

\item{lesson}{lesson name}

\item{student}{student's email}

\item{instructor}{instructor's name}
}
\description{
Run this to set up a SocraticSwirl lesson. Particular exercises can then
be accessed using the \code{\link{exercise}} function.
}

