% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{socratic_swirl_acl}
\alias{socratic_swirl_acl}
\title{Create an ACL (Access Control List) object for instructor-only objects}
\usage{
socratic_swirl_acl()
}
\description{
Create an ACL preventing anyone but the instructor from seeing the student's
response.
}

