% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{socratic_swirl_console}
\alias{socratic_swirl_console}
\title{start up an instructor window}
\usage{
socratic_swirl_console(lesson, course = "none")
}
\description{
Start an instructor window in a Shiny app.
}

