% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{socratic_swirl_error}
\alias{socratic_swirl_error}
\title{Function called after an error during a SocraticSwirl attempt}
\usage{
socratic_swirl_error()
}
\description{
Function called after an error during a SocraticSwirl attempt
}

