% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{socratic_swirl_init}
\alias{socratic_swirl_init}
\title{Initiation of socraticswirl}
\usage{
socratic_swirl_init(id, app, api, instructor, instance = "prod")
}
\arguments{
\item{id}{student's email}

\item{app}{app key}

\item{api}{api key}

\item{instructor}{instructor's name}
}
\description{
Run this to validate a student to use the software for a class. Only once is needed,
before using the \code{\link{socratic_swirl}} function.
}

