% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{socratic_swirl_options}
\alias{socratic_swirl_options}
\title{extract socratic_swirl options from environment}
\usage{
socratic_swirl_options(error = TRUE)
}
\arguments{
\item{error}{whether to raise an error}
}
\value{
A list containing
  \item{course}{course name for current SocraticSwirl session}
  \item{lesson}{lesson name for current SocraticSwirl session}
}
\description{
extract socratic_swirl options from environment
}

