% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{start}
\alias{start}
\title{Take an instructor-provided lesson with SocraticSwirl}
\usage{
start()
}
\arguments{
\item{exercise}{Which quiz exercise to take; provided by instructor}
}
\description{
This is to be called after \code{\link{socratic_swirl}} is used to set up
a SocraticSwirl session.
}

