% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{upload_exercises}
\alias{upload_exercises}
\title{For an instructor; upload exercises to the SocraticSwirl database}
\usage{
upload_exercises(input)
}
\arguments{
\item{input}{Either a single .yaml file, or a directory containing...}
}
\description{
For an instructor; upload exercises to the SocraticSwirl database
}

