% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/socraticswirl.R
\name{exercise}
\alias{exercise}
\title{Take an instructor-provided exercise with SocraticSwirl}
\usage{
exercise(exercise)
}
\arguments{
\item{exercise}{Which quiz exercise to take; provided by instructor}
}
\description{
This is to be called after \code{\link{socratic_swirl}} is used to set up
a SocraticSwirl session.
}

