% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/api.R
\name{socratic_swirl}
\alias{socratic_swirl}
\title{set up SocraticSwirl in this session}
\usage{
socratic_swirl(lesson, instructor, course = "none", ...)
}
\arguments{
\item{lesson}{lesson name}

\item{instructor}{instructor's name}

\item{course}{course name, default \code{"none"}}

\item{...}{extra arguments, not yet used}
}
\description{
Run this to set up a SocraticSwirl lesson. Particular exercises can then
be accessed using the \code{\link{exercise}} function.
}

